Create Code.cpp
We have proven that no two neighboring philosophers can eat at the same time using the above solution to the dining philosopher problem. The disadvantage of the above approach is that it may result in a deadlock situation. This occurs when all of the philosophers pick their left fork at the same time, resulting in a deadlock situation in which none of the philosophers can eat.